src/copilot_sdk_flow/requirements.txt
# those are the dependencies required only by chat.py

# openai SDK
openai==1.13.3

# promptflow packages
promptflow==1.9.0
promptflow-tracing==1.9.0
promptflow-tools==1.4.0

# azure dependencies (for authentication)
azure-core==1.30.1
azure-identity==1.16.0

# utilities
pydantic>=2.6
